services/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
boto3>=1.34.0
pydub>=0.25.1
audioop-lts>=0.2.1; python_version >= "3.13"
python-multipart>=0.0.9
Pillow>=10.0.0
pytesseract>=0.3.10
